feat: tabs use store
