+++ src/client/components/NavGroup.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div flex items-center gap-12px>
+    <VButton>
+      <div i-material-symbols-ungroup />
+      Reset group
+    </VButton>
+  </div>
+</template>
